spark quarantine
